+++ src/main/java/kitchenpos/domain/menu/MenuRepository.java
@@ -1,7 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.domain.menu;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
