+++ src/main/java/kitchenpos/dto/menu/MenuProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.menu;
 
 public class MenuProductRequest {
