+++ src/main/java/kitchenpos/dto/menu/MenuResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.dto;
+package kitchenpos.dto.menu;
 
 import static java.util.stream.Collectors.toList;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 
 public class MenuResponse {
